requirements.txt
asgiref==3.5.2
cloudinary==1.29.0
Django==3.2.15
django-cloudinary-storage==0.3.0
django-filter==22.1
django-location-field==2.1.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
